+++ src/Plugin.php
@@ -124,20 +124,6 @@ class Plugin extends \craft\base\Plugin
     }
 
     /**
-     * @inheritdoc
-     */
-    public function beforeUpdate(string $fromVersion): bool
-    {
-        if (version_compare($fromVersion, '1.1.0', '<')) {
-            Craft::error('Social Login 2 requires you to be running at least v1.1.0 before updating');
-
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
      * Save plugin settings.
      *
      * @param array $settings